# api_docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EIT-P API Documentation",
    "description": "EIT-P (Emergent Intelligence through Thermodynamic Principles) API文档",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8085",
      "description": "API服务器"
    },
    {
      "url": "http://localhost:8086",
      "description": "推理服务"
    },
    {
      "url": "http://localhost:8087",
      "description": "认证服务"
    },
    {
      "url": "http://localhost:8082",
      "description": "监控仪表板"
    }
  ],
  "paths": {
...
    "/api/inference/text_generation": {
      "post": {
        "summary": "文本生成推理",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": {"type": "string"},
                  "model_id": {"type": "string"},
                  "max_length": {"type": "integer"},
                  "temperature": {"type": "number"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文本生成成功"
          }
        }
      }
    },
...
  }
}